modif config generator pour inclure a main. modif script envoi/reception code ceci

+++ launch_slurm_jobs.sh
@@ -4,13 +4,17 @@ script_dir=$(dirname "$0")
 slurm_dir="$script_dir/slurm"
 master_dir="$slurm_dir/master"
 
-rm -rf $slurm_dir
-rm -rf "$script_dir/plots"
-rm -rf "$script_dir/results"
-rm -rf "$script_dir/output"
-rm "$master_dir/master.slurm"
+[ -d "$slurm_dir" ] && rm -rf "$slurm_dir"
+[ -d "$script_dir/plots" ] && rm -rf "$script_dir/plots"
+[ -d "$script_dir/results" ] && rm -rf "$script_dir/results"
+[ -d "$script_dir/output" ] && rm -rf "$script_dir/output"
 
-python3 "$script_dir"/main.py generate_configs
-python3 "$script_dir"/main.py generate_slurm_files
+python3 "$script_dir/main.py" generate_configs || { echo "Échec de la génération des configurations"; exit 1; }
+python3 "$script_dir/main.py" generate_slurm_files || { echo "Échec de la génération des fichiers Slurm"; exit 1; }
 
-sbatch "$master_dir"/master.slurm 
+if [ -f "$master_dir/master.slurm" ]; then
+    sbatch "$master_dir/master.slurm" || { echo "Échec de la soumission du job master.slurm"; exit 1; }
+else
+    echo "Erreur: $master_dir/master.slurm n'existe pas"
+    exit 1
+fi
